backend/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
Brlapi==0.8.2
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
chrome-gnome-shell==0.0.0
cryptography==37.0.4
cupshelpers==1.0
dbus-python==1.2.16
distro==1.5.0
distro-info==1.0
Django==4.0.6
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
feedparser==5.2.1
haversine==2.5.1
httplib2==0.18.1
idna==2.10
Jinja2==3.1.2
jwt==1.3.1
locationiq==0.0.2
louis==3.16.0
Mako==1.1.3
MarkupSafe==2.1.1
musicbrainzngs==0.7.1
mutagen==1.45.1
numpy==1.22.0
openpyxl==3.0.9
pandas==1.3.5
pep8==1.7.1
powerline-status==2.8.1
psutil==5.8.0
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.21
pycups==2.0.1
pycurl==7.43.0.6
PyGObject==3.38.0
pyinotify==0.9.6
PyJWT==2.4.0
pyperclip==1.8.2
PySimpleSOAP==1.16.2
pysmbc==1.0.23
python-apt==2.2.1
python-dateutil==2.8.2
python-debian==0.1.39
python-debianbts==3.1.0
python-http-client==3.3.7
pytz==2021.3
pyxdg==0.27
quodlibet==4.3.0
reportbug==7.10.3+deb11u1
requests==2.25.1
schedule==1.1.0
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
toml==0.10.2
trash-cli==0.17.1.14
unattended-upgrades==0.1
urllib3==1.26.5
xdg==5
XlsxWriter==3.0.3
